Commit 'custom_targets.json'.
This is the interesting bit and what I've been building up to.

Mbed OS 5 program won't build if the target is 'DISCO_L053C8' because the target definition contains '"release_versions": ["2"]'.

It turns out it is possible to define a custom target and override 'release_versions' so that it will build. It appears to run as well, although it doesn't behave identically to an mbed OS 2 equivalent program.