Simple 'EventQueue' investigation.
https://os.mbed.com/blog/entry/Simplify-your-code-with-mbed-events/ does a good sales job on mbed events. Also the LoRaWAN example uses an 'EventQueue'.

Hence I thought I'd do some investigation. It turns out 'EventQueue' is not part of the RTOS hence this examples works fine.

'EventQueue' is dynamically allocating the memory for the queue but it happens in the constructor which is run before 'main' and hence there is not trace for it.