Document sleep tracing macro.
https://os.mbed.com/docs/v5.8/reference/power-management.html describes sleep and deep sleep. I'm very interested in low power modes.

One of the things described is the "Sleep/deep sleep profiling tool". This is enabled with the MBED_SLEEP_TRACING_ENABLED macro.

The profiling tool produces a lot of output so I don't want it enabled all the time but I did want to document its existance for myself.

With the profiling tool enabled I could see that there are always active sleep locks preventing deep sleep. Which I think is because of the event library default config.

I've discovered that json doesn't have comments and therefore the easiest way to /comment/ something is to change the field name slightly so that is is just ignored.