Use low power timer for events.
Low power is something we're interested in so I wanted to investigate this option.

Using the low power timer means that the sleep manager reports that it can deep sleep. I've no idea if it is actually selecting deep sleep or how much power it saves.

It took me a while to understand how to override the default events library configuration but I think I get it now.